removed 1.6-only code from WorkbenchPaneSS.java
fixed csv and xls workbench import issues.

git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@1598 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/tasks/subpane/wb/CSVImport.java
@@ -84,7 +84,7 @@ public class CSVImport implements DataImport
                 for (int col = 0; col < csv.getColumnCount(); col++)
                 {
                     // Skip the column if it isn't found in the hash
-                    WorkbenchTemplateMappingItem wbtmi = colHash.get(col);
+                    WorkbenchTemplateMappingItem wbtmi = colHash.get((short) col);
                     if (wbtmi != null)
                     {
                         wbRow.setData(csv.get(col), wbtmi.getViewOrder());
